feat: backward compatibility with tutor14
